Updating scales2.h

+++ AnalysisStep/macro/scales2.h
@@ -185,7 +185,7 @@ float getGGZZQCDScaleUncertainty7TeV(float m4l) {
 
 ///////////// Pileup reweighting /////////////////////////////
 
-float getPUWeight(int numsim, int mode=1) {
+float getPUWeight(float numsim, int mode=1) {
     return is2011 ? getPUWeight2011(numsim) : getPUWeight2012(numsim, mode);
 }
